+++ monkey_web/recipes/rm.rb
@@ -1,9 +1,11 @@
-execute 'remove containers' do
-    only_if { Dir.exist?("/srv/monkey_web/") }
-    cwd '/srv/monkey_web/'
-    command 'docker-compose rm --force'
-    case node[:platform]
-    when 'ubuntu'
-      environment 'COMPOSE_API_VERSION' => '1.18'
-    end
+if File.exist?('/srv/monkey_web/docker-compose.yml')
+  execute 'remove containers' do
+      only_if { Dir.exist?("/srv/monkey_web/") }
+      cwd '/srv/monkey_web/'
+      command 'docker-compose rm --force'
+      case node[:platform]
+      when 'ubuntu'
+        environment 'COMPOSE_API_VERSION' => '1.18'
+      end
+  end
 end
